refactor: Slot 패키지 이동
