+++ src/test/java/roomescape/slot/domain/ReservationSlotTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.domain;
+package roomescape.slot.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
